requirements.txt
# ── Core ──────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Data & Signal Processing ──────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.2
opencv-python>=4.8.0
Pillow>=10.1.0
mne>=1.6.0                    # EEG processing (DEAP/SEED reader)

# ── Deep Learning Extras ──────────────────────────────────────
timm>=0.9.12                  # pretrained CNN backbones
einops>=0.7.0                 # tensor ops for attention

# ── Explainability ────────────────────────────────────────────
shap>=0.44.0
captum>=0.7.0                 # Grad-CAM, integrated gradients

# ── Hyperparameter Tuning ─────────────────────────────────────
optuna>=3.4.0
optuna-dashboard>=0.14.0

# ── Visualisation ─────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ── Config & Logging ─────────────────────────────────────────
pyyaml>=6.0.1
tqdm>=4.66.0
tensorboard>=2.15.0
loguru>=0.7.2

# ── Deployment ────────────────────────────────────────────────
streamlit>=1.29.0
streamlit-webrtc>=0.47.0      # webcam stream
av>=10.0.0                    # video frames

# ── Misc ──────────────────────────────────────────────────────
joblib>=1.3.2
omegaconf>=2.3.0
